+++ src/zope/testrunner/options.py
@@ -604,7 +604,6 @@ def get_options(args=None, defaults=None):
     if options.showversion:
         dist = pkg_resources.require('zope.testrunner')[0]
         print('zope.testrunner version %s' % dist.version)
-        options.fail = True
         return options
 
     if options.subunit or options.subunit_v2: